[InstructionRNN] Dataset returns strokes before annotation as well
Currently not being used in model yet.

Note on using them:
- With text inputs, we could use a separator token between the pre_strokes and strokes.
However, the stroke inputs don't lend themselves to a separator token.
- With a RNN, we could embed each separately and then add the final hidden representations.
- However, the Transformer attends over all encoded hidden states, and there's no good way to
combine them given that pre_strokes and strokes may be of different lengths.
- We could concatenate the strokes and pre_strokes. This is cumbersome because there would
be padding elements between the two. This would require refactoring the padding_masks for
the transformer (i.e. it takes a strokes_lens and pre_strokes_lens).
- We could also just have the Dataset return the combined strokes and prestrokes (initializing
the dataset with a use_pre_strokes flag), which would circumvent the above padding problem.