[InstructionRNN] Flag to decode with z at every time step
Motivation: SketchRNN does this.

Prelim results seem to indicate that this doesn't affect the
loss.... which possibly aligns with other preliminary results
using both rnn and transformers. The encoded strokes seems
to be sort of ignored. Maybe we need to pretrain an
autoencoder on the strokes to get better representations.

Test plan:
--condition_on_hc True
--condition_on_hc False