[Instruction] Encode prestrokes, StrokeEncoderLSTM
1) Create StrokeEncoderLSTM
2) Prestrokes are strokes that preceded annotated strokes.
There's a couple possible ways to use them.
Currently, this encodes the prestrokes, then (using the
hidden and cell states), encodes the strokes.
- Perhaps it would be useful to have a larger dependence on
the prestrokes. We could sum / concatenate the hidden
and cell with the pre_hidden and pre_cell.
- We could also,similar to condition_on_hc, separately condition on the
pre_hidden and pre_cell in the InstructionDecoderLSTM

Test plan:
-- use_prestrokes True