[InstructionGen] First pass at calculating rare words stat
Motivation:
Maximum likelihood training often reduces in shorter, more generic
generations. This is a known problem in dialogue generation, translation,
etc. We want to measure the distribution of rare words being produced.

This first pass simply counts the number of unique words in the
reference vs. generated instructions. Results with the
cnn-encoder + lstm-decoder + decoding with categories:
Number of unique tokens in reference instructions:  188
Number of unique tokens in generated instructions:  58