[Instruction] Concate prestrokes with strokes with a separator point
Previously, was either:
a) in the LSTM encoder case, encoding prestrokes and initializing lstm for strokes
with prestrokes' hidden and cell
b) in the transformer encoder case, encoding presrokes and strokes separately

Now, we create a separator point. This is [0,0,1,1,0], which never occurs otherwise in the
dataset as pt[2] and pt[3] are pendown and penup.

Results:
Model(s) seem to do a little worse with prestrokes actually

Test plan:
Example:
- CUDA_VISIBLE_DEVICES=7 PYTHONPATH=. python src/models/instruction_gen.py --notes prestrokes-concat --model_type lstm

Use --model_type 'lstm', 'transformer_lstm', 'cnn_lstm', and also tested without prestrokes (--use_prestrokes false)