[Refactor] Add comments / clean up sketch generation code
Still only generates one item per batch. Added comments
for input outputs.

NOTE: I also added an adjust_temp when calculating the q.
This isn't done in the magenta repo, but it seems to help.

Note: also added a nan set_trace() because it seems like
there is very occassionally a cuda device error that gets thrown
at the binary cross entropy loss between p and q. Apparently,
this happens if q is not within [0,1]. This shouldn't happen
given that q is calculated using a softmax. There must be a
nan occurring after the softmax.
- Given that there was also very occassionally a nan in the
dx dy reconstruction loss (LS), it seems like the error
is not specific to either loss calculation; rather there
is probably a common bug with the decoder generating those
tensors.