[Segmentation] Hackily save progressions and segmentations
Goal: want to save a) the segmentations, and b) the progression
of the drawing (each segment corresponds to one penup) so that
we can see if the generated subinstructions and segmentations
make sense.

This commit does a super hacky probably broken implementation.
The main problem is converting from the stroke5 data from the
StrokeDataset (originally derived from the SketchRNN npz data)
to the ndjson format so that I could reuse the progression plotting
data I already implemented earlier.
- I need to check what is the correct way to unnormalize,
convert from delta-x/y's to absolute x/y's, position the x and y's
(no negative numbers?), and scale them.

Right now, I hacked it so that the drawings show up within the frame.

Test plan:
- PYTHONPATH=. python src/models/segmentation.py