[Segmentation][WIP] Use ProgressionPairDataset
Motivation: Think there might be a problem with normalizing
the StrokeDataset for use with a trained Stroke2Instruction
model (which is trained on the ProgressionPairDataset).
So let's run the S2I model on the ProgressionPair dataset.

Had to modify PP dataset to take a flag so that it returns
the full stroke3 drawing and not just the annotated segment.

WIP: The drawings look totally incorrect