[SketchRNN] Train on annotated (progressionpair) dataset
Motivation: this is a baseline (no instructions).

Had to create a collate_fn that only returns strokes and category
information for ProgressionPair dataset. This way SketchRNN
can use either the ndjson dataset or the progressionpair dataset.
Current design is not great... batch is unpacked throughout
SketchRNN, and what is contained in a batch may vary.

Test plan:
- PYTHONPATH=. python src/models/sweeps/sketch_rnn_sweep.py --groupname annotated_dataset_baseline