[App] Show two instruction trees for each image
Motivation: testing new ways to generate instruction trees,
want to qualitatively assess the differences

Test plan:
1. gunicorn -w=2 -b=0.0.0.0:8080 --chdir=app wsgi:application --preload
2. gunicorn -w=2 -b=0.0.0.0:8080 --chdir=app wsgi:application --preload --daemon

TODO: the above involves killing gunicorn and restarting it every time
to test... not the proper way to do it. I should be able to do the
normal PYTHONPATH=. python app/application.py, but there is a path error
with static now since I refactored to allow gunicorn to work.