[Data] Save drawing segs as images for S2I
Motivation: Stroke to instruction model is hard to learn
for a number of reasons (not a lot of data, drawings are
noisy, stroke data is complicated). To address the last point,
let's try encoding the drawings as images.

For every drawing, divide it up into beginning (pre),
middle (current), and end (post) segments. Save each one of these
as a drawing, and also save the final drawing. The encoder can
take as input an "image" with each of these as a channel (i.e.
so that it sees the "current" portion that is annotated, the
portion of the drawing before and after, and the final drawing).

Test plan:
PYTHONPATH=. python src/data_manager/quickdraw.py --precurrentpost