[S2I][Images] Add diferent types of CNNs
Motivation: the channels that compose an image are a
bit unique. We would probably expect that the
annotated segment to be the most important, with some
dependencies on the other channels (pre, full, etc.).

To model thise channel-wise relationships:

1. Squeeze and Excitation (channel-wise)

Resources:
- https://github.com/moskomule/senet.pytorch
- https://arxiv.org/abs/1709.01507

2. CBAM  (channel-wise crossed with spatial)

Resources:
- https://github.com/Jongchan/attention-module
- http://openaccess.thecvf.com/content_ECCV_2018/papers/Sanghyun_Woo_Convolutional_Block_Attention_ECCV_2018_paper.pdf

Test plan:
- Ran sweep with S and E
- CBAM not working yet because it expects images
in either the typical CIFAR size or ImageNet size
(ours are 112 x 112)