[Data][Bugfix] Stroke lens off by one
In quickdraw.py's stroke3_to_stroke5, see this line:
result[l:, 4] = 1  # last "stroke" has stroke5[4] equal to 1, all other values 0 (see Figure 4); hence l
where l is equal to the stroke_len.
This last point is the end of drawing point. By using l (stroke_len)
as the stroke_lens, I was off by one and not including this last point.

Not sure what the effect of that is going to be on all the models...
the sketchrnn models were ending though, so I'm not sure how it was
learning that.