Add script that investigates difference between npz and ndjson
Motivation: I have been able to reproduce the unconditional generation
model's loss numbers (SketchRNN decoder only) on the pig category when
training on the npz dataset (originally used in ndjson). However,
the loss numbers are much higher (0.1 vs -0.6) when training on the ndjson
dataset. This investigates that.

Turns out, the ndjson dataset is a simplified version of the npz drawings, as
the average length is shorter for ndjson than npz. This despite the
website saying that they both use (RDP with epsilon=2 to simplify).
https://en.wikipedia.org/wiki/Ramer%E2%80%93Douglas%E2%80%93Peucker_algorithm.
(I had originally used the ndjson dataset because it contains extra metadata,
such as whether or not the drawing was recognized by a trained classifier
during QuickDraw.)

Nevertheless, the generated drawings for a npz model vs a ndjson model seem
similar though.... If anything the ndjson is slightly better? I'm not sure
why the loss numbers vary so much though. It's okay.