[Plan] Leaves only run; [Segment] sweep
1. Map entire drawing's hidden states to the concatenated
leaf instructions. (I.e. same as toplevel-leaves but with just
the leaves).

(Next commit will do stack-leaves (map each segment's hidden state
to that segment's leaf instruction).)

Note: this is still without layer norm, but the previous (toplevel
and toplevel-leaves) runs I'm comparing against the SketchRNN baseline
are also without layernorm.

Test plan:
PYTHONPATH=. python src/models/sweeps/sketch_with_plans_sweep.py --groupname leaves

2. Segment sweep is simply finishing running the segmentation sweep
that is the BEST_NDJSON_SEGMENTATION... in config.py. This was
Feb20 and actually is the old LSTM model from bigsweep, but the probabilities
are corrected (so that they're probabilities and not logits). I had previously
run it on Feb20 for just ~2000 per category. Hardcoded it so that
it saved into the Feb20 run (even though I ran this on Mar20).