[VAEzToInstruction] Add first pass of model
Motivation: before / in parallel with doing the
Instruction to VAEz model, which can then be used to
improve the sketchwplans model, let's look at if we can
learn z -> instructions. To what degree does the latent vector
contain the information for the sequence of the drawing (so
much as the instructions contain it)?

Test plan:
See that it runs:
CUDA_VISIBLE_DEVICES=5 PYTHONPATH=. python src/models/vaez_to_instruction.py --categories pig --max_per_category 128