[SketchWPlans][Z] Re-add cond_instructions=='decinputs' option
Motivation / background:
- I had tested cond_instructions=='decinit' and 'decinputs' previously.
- The latter encodes the instructions, and passes that representation
to the Sketch decoder at every time step (similar to how the
SketchRNNVAE passes the z to every time step).
- These did not help the model (loss was equal to the baseline).
- Thus, I switched to the coond_instructions=='match', where we
shape the decoder representations by applying a triplet loss between
them and the encoded instruction representations.

- Now, I've been working on learning instruction to VAEz mappings.
We know the z's help the sketch model. So if we can learn a better
mapping (instead of doing it from scratch as before), it should help.
So let's add decinputs back in.